Refactor team access code
