+++ src/app/(dashboard)/[team_slug]/[workspace_slug]/layout.tsx
@@ -6,8 +6,9 @@ import { ReactNode } from "react";
 export default function WorkspaceLayout({ children }: { children: ReactNode }) {
   const param = useParams() as { team_slug: string; workspace_slug: string };
   const pathname = usePathname();
-  return pathname ===
-    `/${param?.team_slug}/${param?.workspace_slug}/settings` ? (
+  return pathname?.includes(
+    `/${param?.team_slug}/${param?.workspace_slug}/settings`,
+  ) ? (
     children
   ) : (
     <WorkspaceView>{children}</WorkspaceView>
