+++ src/app/api/teams/[team_slug]/invites/reset/route.ts
@@ -1,11 +1,11 @@
-import { withAuth } from "@/lib/auth";
+import { withTeam } from "@/lib/auth";
 import mongoDb, { databaseName } from "@/lib/mongodb";
 import { randomId } from "@/lib/utils";
 import { ObjectId } from "mongodb";
 import { NextResponse } from "next/server";
 
 // POST /api/teams/[team_slug]/invites/reset – reset invite code for a team
-export const POST = withAuth(
+export const POST = withTeam(
   async ({ team }) => {
     const client = await mongoDb;
     const teamsDb = client.db(databaseName).collection("teams");
